backend/requirements.txt
APScheduler==3.10.4
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
PyJWT==2.12.1
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2026.1.post1
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
Werkzeug==3.1.7
gunicorn==23.0.0
pandas==2.2.3
openpyxl==3.1.5
reportlab==4.2.5
razorpay==1.4.2